refactor: 모든 SQL문을 대문자로 수정
