+++ src/main/resources/schema.sql
@@ -1,24 +1,23 @@
-create table if not exists STATION
+CREATE TABLE IF NOT EXISTS STATION
 (
-    id bigint auto_increment not null,
-    name varchar(255) not null unique,
-    primary key(id)
+    id BIGINT AUTO_INCREMENT NOT NULL,
+    name VARCHAR(255) NOT NULL UNIQUE,
+    PRIMARY KEY(id)
 );
 
-create table if not exists LINE
+CREATE TABLE IF NOT EXISTS LINE
 (
-    id bigint auto_increment not null,
-    name varchar(255) not null unique,
-    color varchar(20) not null,
-    primary key(id)
+    id BIGINT AUTO_INCREMENT NOT NULL,
+    name VARCHAR(255) NOT NULL UNIQUE,
+    color VARCHAR(20) NOT NULL,
+    PRIMARY KEY(id)
 );
 
-create table if not exists SECTION
+CREATE TABLE IF NOT EXISTS SECTION
 (
-    id bigint auto_increment not null,
-    line_id bigint not null,
-    up_station_id bigint not null,
-    down_station_id bigint not null,
-    distance int,
-    primary key(id)
+    id BIGINT AUTO_INCREMENT NOT NULL,
+    up_station_id BIGINT NOT NULL,
+    down_station_id BIGINT NOT NULL,
+    distance int NOT NULL,
+    PRIMARY KEY(id)
 );
